react,next upgrade - alpha 4
