+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "start": "next start",
+    "start": "PORT=3000 next start",
     "lint": "npm run lint:code && npm run lint:styles",
     "lint:fix": "npm run lint:code:fix && npm run lint:styles:fix",
     "lint:code": "eslint ./src --ext .js,.jsx,.ts,.tsx",
@@ -21,10 +21,10 @@
   "dependencies": {
     "camelcase": "^6.2.0",
     "framer-motion": "2.8.0",
-    "next": "10.0.0",
-    "react": "16.13.1",
+    "next": "^11.0.1",
+    "react": "^17.0.2",
     "react-copy-to-clipboard": "^5.0.3",
-    "react-dom": "16.13.1",
+    "react-dom": "^17.0.2",
     "react-infinite-scroll-component": "^6.1.0",
     "react-intersection-observer": "^8.32.0",
     "react-markdown": "4.3.1",
